// src/modules/ProfileMgmt/types.tsx
export type User = {
  id?: string;
  firstName: string;
  lastName: string;
  fullName?: string;
  gender?: string;
  dateOfBirth: string;
  nationality: string;
  address?: string;
  phoneNumber: string;
  email: string;
  skype?: string;
  position: string;
  employmentDate: string;
  lengthOfService?: number;
  grade: number;
  description?: string;
  avatar?: string;
};